pyproject.toml
[project]
name = "crontab-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "python-crontab>=3.2.0",
    "reflex>=0.7.3",
    "ruff>=0.11.2",
]
